[fix] heap after use
reason of error was lifetime of user.

If the client dies while the server is sending the message, a Broken Pipe error is thrown, and the user is reset via on_inactive as part of the handling of that error.

It then returns for the rest of the transfer, referencing the user object that was just reset.

Fixed: Increased the lifetime of user objects: on_inactive now does not reset the user.